Fix for id 00 issue unfoldingWord-dev/Door43#300
This commit fixes the last ID of each chapter which previously showed up
as something like 29-00.  We do not want to push this into the v2 API as
this will mess up existing tS users.  We *do* want to incorporate this
into the tD publishing as that will become the v3 api.